dont permit accented characters in package names
to ensure url and human friendly names if the user enters accented characters as required for npm modules

ttps://npmjs.org/package/normalize-package-data

Rules for name field

If name field is given, the value of the name field must be a string. The string may not:

start with a period.
contain the following characters: /@\s+%
contain and characters that would need to be encoded for use in urls.
resemble the word node_modules or favicon.ico (case doesn't matter).