Update package.json
Updated q-io dependency to a version that uses copyTree whenever an EXDEV error occurs in move. 
EXDEV is a common error when running minit in FreeBSD (and probably other *nixes) since rename(2) isn't allowed if the link named by "to" and the file named by "from" are on different logical devices (file systems). Which is the case of /tmp.