Manually set the npm prefix before running install
I had tried all the other configuration setting approaches yesterday to
no avail, and I thought I'd checked this one. But either way, this is
a lot easier than forking a new npm process, especially when we want to definitely rely on the npm packaged as part of minit.

It is troubling that any other application that has multiple libraries
working with npm might need to step in to coordinate them from trashing
each other's configurations.