Return the final path of the created result
Prior to this it fell to the consumer of minit to build the final path
and the extension was not present anywhere. Now this information is
ferried along in the promise resolution.

It's debatable about what should actually be returned from the creation
API. It feels a little weird adding output into the input config.